portfolio-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic>=2.9.0
